fix incorrect build tags && create common file for all platforms
