+++ libbeat/cmd/instance/metrics_other.go
@@ -1,20 +1,12 @@
-// +build !darwin,!linux,!cgo darwin,!cgo freebsd,!cgo openbsd
+// +build !darwin !cgo
+// +build !freebsd !cgo
+// +build !linux,!windows
 
 package instance
 
 import (
-	"github.com/satori/go.uuid"
-
 	"github.com/elastic/beats/libbeat/logp"
 )
-
-var (
-	ephemeralID uuid.UUID
-)
-
-func init() {
-	ephemeralID = uuid.NewV4()
-}
 
 func setupMetrics(name string) error {
 	logp.Warn("Metrics not implemented for this OS.")
